# Optimization: concat function