## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes https://github.com/apache/arrow-datafusion/issues/9742

## Rationale for this change

optimize `concat` and `concat_ws` function.

### Benchmark(only concat)

```shell
Gnuplot not found, using plotters backend
concat function/concat(old)/1024
                        time:   [91.562 µs 91.725 µs 91.905 µs]
Found 8 outliers among 100 measurements (8.00%)
  6 (6.00%) high mild
  2 (2.00%) high severe
concat function/concat(new)/1024
                        time:   [17.831 µs 17.877 µs 17.934 µs]
Found 9 outliers among 100 measurements (9.00%)
  7 (7.00%) high mild
  2 (2.00%) high severe

concat function/concat(old)/4096
                        time:   [357.95 µs 358.98 µs 360.02 µs]
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe
concat function/concat(new)/4096
                        time:   [71.003 µs 71.168 µs 71.326 µs]
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

concat function/concat(old)/8192
                        time:   [758.82 µs 761.43 µs 764.82 µs]
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
concat function/concat(new)/8192
                        time:   [141.11 µs 141.44 µs 141.79 µs]

```

### Attention

For the purpose of benchmarking, I haven't officially replaced the `concat` and `concat_ws` function yet. If the community finds this PR meaningful, I will proceed with the replacement.

## What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

## Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

## Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->