@alamb 2024-03-22
I also filed https://github.com/apache/arrow-datafusion/issues/9742 to track this improvement